Move ConfirmationDialogView to *.AndroidS.AppCompat.App
