+++ HoneyComb.Platform.UI.AndroidS/AppCompat/App/ConfirmationAndroidViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using Android.Arch.Lifecycle;
 using HoneyComb.UI.Dialogs;
 
-namespace HoneyComb.Platform.UI.AndroidS
+namespace HoneyComb.Platform.UI.AndroidS.AppCompat.App
 {
     public sealed class ConfirmationAndroidViewModel : ViewModel
     {
